Revert "Refactor into runloop.rs"
This reverts commit f223d9df4b8da9e7e3e34fc706c5ff4ee4210931.
