--- index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Learn WGPU</title>
-    <style>
-        canvas {
-            background-color: black;
-        }
-    </style>
-</head>
-
-<body id="wasm-example">
-  <script type="module">
-      import init, { run } from "./pkg/minecrust.js";
-      init().then(() => {
-          console.log("WASM Loaded");
-          run();
-      });
-  </script>
-</body>
